Moves printers from cmd to pkg
