+++ pkg/printers/table/collector.go
