+++ pkg/printers/table/collector_test.go
